+++ docs/History.md
@@ -2,6 +2,10 @@
 
 #### TransektCount Copyright (C) 2016-2019, Wilhelm Stein
 
+#### Version 3.1.0 ()2019-04-16)
+- Integrated list of many European species
+- Instead of manual input, add further species to your section lists by scroll-down selection from a comprehensive list that contains only species not contained yet
+
 #### Version 3.0.8 (2019-04-04)
 - Butterfly icons added to section editor